Weaken the superclass on `Commutative`
That it was something as strong as `Fractional` was a relic of the `Group` superclass to `Abelian` I stripped away to make this package.

However, `Num`'s suggested laws do not come with commutativity for
`(*)`. Should we do this anyways? or should we have a
```haskell
class Num a => CommutativeMultiplication a
```
helper class so we can do
```haskell
instance CommutativeMultiplication a => Commutative (Product a)
```
?